Auth and Trips Controllers
Moved auth and trips controllers inside their own files
